v2.0.6.12
여제가 좌표를 융통성있게 판단하도록 변경
조사처리 관련 수정

+++ src/daybreak/abilitywar/ability/list/TheEmpress.java
@@ -66,8 +66,8 @@ public class TheEmpress extends AbilityBase {
 				if(!Cool.isCooldown()) {
 					Location l = getPlayer().getLocation();
 					
-					NumberStatus X = NumberUtil.getNumberStatus(l.getX());
-					NumberStatus Z = NumberUtil.getNumberStatus(l.getZ());
+					NumberStatus X = NumberUtil.getNumberStatus((int) l.getX());
+					NumberStatus Z = NumberUtil.getNumberStatus((int) l.getZ());
 					
 					Random random = new Random();
 					boolean bool = random.nextBoolean();
